Fixed Tempo type (int -> double)

+++ src/als-tools.core/Entities/LiveProject.cs
@@ -53,7 +53,7 @@ public class LiveProject
     /// <summary>
     /// The project tempo/BPM
     /// </summary>
-    public int Tempo { get; set; }
+    public double Tempo { get; set; }
 
     /// <summary>
     /// The project time signature. It is probably a bit mask.
